Move card and slider into coupons directory.
