+++ plugins/woocommerce-admin/client/marketing/components/index.js
@@ -1,6 +1,4 @@
-export { default as Card } from './card';
 export { default as ProductIcon } from './product-icon';
-export { default as Slider } from './slider';
 export { default as ReadBlogMessage } from './ReadBlogMessage';
 export { CollapsibleCard, CardBody, CardDivider } from './CollapsibleCard';
 export { PluginCardBody, SmartPluginCardBody } from './PluginCardBody';
