+++ plugins/woocommerce-admin/client/marketing/coupons/card/style.scss
