+++ plugins/woocommerce-admin/client/marketing/coupons/recommended-extensions/index.js
@@ -13,8 +13,8 @@ import PropTypes from 'prop-types';
 import './style.scss';
 import RecommendedExtensionsItem from './item';
 import RecommendedExtensionsPlaceholder from './placeholder';
-import { STORE_KEY } from '../../data/constants';
-import Card from '../../components/card';
+import { STORE_KEY } from '~/marketing/data/constants';
+import Card from '../card';
 
 const RecommendedExtensions = ( {
 	extensions,
