+++ plugins/woocommerce-admin/client/marketing/coupons/slider/index.js
