+++ plugins/woocommerce-admin/client/marketing/coupons/slider/style.scss
